Remove uses of executor handles, rename mod tokio -> future
